test(org): split ax.test.ts (1549 lines) into per-concern files
Follow-up to #421's exemplar, applied to the largest test file. The four
biggest cohesive sections move into focused files sharing one fixture
module:
  - _axHelpers.ts            — shared bootstrap + runGate + rid/today
  - axSuggest.test.ts        — gate suggest tight-loop
  - axVerbsAvailableNow.test.ts — boot.verbs_available_now discovery
  - axVoiceCalibration.test.ts  — Two-Persona Devil voice memory
  - axThank.test.ts          — gate thank primitive
ax.test.ts (1549 → 648 lines) keeps the remaining AX affordances
(suggested_next routing, JSON error envelope, board/show surfaces,
advisory semantics, transcript, --plain, --dry-run, text footer).

Same 56 tests (verified by count + a focused run of all five files), no
behavior change. All blocks were contiguous cuts at the `// ──` section
boundaries and depend only on the extracted helpers. boot.test.ts is the
remaining large file.

Full suite 1790 pass / 0 fail.
