+++ tests/interface/ax.test.ts
@@ -15,61 +15,7 @@
 
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import { spawnSync } from 'node:child_process';
-import {
-  mkdtempSync,
-  writeFileSync,
-  rmSync,
-  mkdirSync,
-} from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join, resolve, dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-const here = dirname(fileURLToPath(import.meta.url));
-const GATE = resolve(here, '../../../bin/gate.mjs');
-
-function bootstrap(): { root: string; cleanup: () => void } {
-  const root = mkdtempSync(join(tmpdir(), 'guild-ax-'));
-  writeFileSync(
-    join(root, 'guild.config.yaml'),
-    'content_root: .\nhost_names: [human]\n',
-  );
-  for (const d of ['members', 'requests', 'issues', 'inbox']) {
-    mkdirSync(join(root, d));
-  }
-  for (const name of ['alice', 'bob']) {
-    writeFileSync(
-      join(root, 'members', `${name}.yaml`),
-      `name: ${name}\ncategory: professional\nactive: true\n`,
-    );
-  }
-  return { root, cleanup: () => rmSync(root, { recursive: true, force: true }) };
-}
-
-function runGate(
-  cwd: string,
-  args: string[],
-  env: Record<string, string> = {},
-  input?: string,
-): { stdout: string; stderr: string; status: number } {
-  const opts: Parameters<typeof spawnSync>[2] = {
-    cwd,
-    env: { ...process.env, ...env },
-    encoding: 'utf8',
-  };
-  if (input !== undefined) opts.input = input;
-  const r = spawnSync(process.execPath, [GATE, ...args], opts);
-  return {
-    stdout: typeof r.stdout === 'string' ? r.stdout : '',
-    stderr: typeof r.stderr === 'string' ? r.stderr : '',
-    status: r.status ?? -1,
-  };
-}
-
-const today = () => new Date().toISOString().slice(0, 10);
-const rid = (n: number) => `${today()}-${String(n).padStart(4, '0')}`;
-
+import { bootstrap, runGate, rid } from './_axHelpers.js';
 // ── boot.suggested_next: workflow-stage routing ───────────────────
 
 test('boot.suggested_next: pending-as-executor → approve', () => {
@@ -269,853 +215,6 @@ test('show --fields: unknown keys silently dropped (not errored)', () => {
     const { stdout } = runGate(root, ['show', rid(1), '--fields', 'state,not_a_field']);
     const payload = JSON.parse(stdout);
     assert.deepEqual(Object.keys(payload), ['state']);
-  } finally {
-    cleanup();
-  }
-});
-
-// ── gate suggest: tight-loop sibling of boot ─────────────────────
-
-test('suggest --format json: returns the same triple as boot, no orientation payload', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const bootOut = JSON.parse(runGate(root, ['boot'], { GUILD_ACTOR: 'bob' }).stdout);
-    const suggestOut = JSON.parse(runGate(root, ['suggest'], { GUILD_ACTOR: 'bob' }).stdout);
-    // Same suggestion.
-    assert.deepEqual(bootOut.suggested_next, suggestOut.suggested_next);
-    // Orientation keys present in boot, absent in suggest.
-    assert.ok('status' in bootOut);
-    assert.ok('tail' in bootOut);
-    assert.equal('status' in suggestOut, false);
-    assert.equal('tail' in suggestOut, false);
-    // Payload shrinks dramatically — suggest is the hot-loop form.
-    const bootStr = JSON.stringify(bootOut);
-    const suggestStr = JSON.stringify(suggestOut);
-    assert.ok(
-      suggestStr.length < bootStr.length / 3,
-      `expected suggest to be <1/3 of boot, got ${suggestStr.length} vs ${bootStr.length}`,
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('suggest: null when registered actor has nothing to do', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    const { stdout } = runGate(root, ['suggest'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.suggested_next, null);
-    // Genuine silence — no open work, no reason to surface.
-    assert.equal(payload.suggested_next_reason, null);
-  } finally {
-    cleanup();
-  }
-});
-
-test('suggest: null with substrate open-work surfaces suggested_next_reason', () => {
-  // Asteria dogfood 2026-05-17: host saw status.pending.total: 1 but
-  // suggested_next: null. The substrate's silence read as a bug. Fix:
-  // a sibling `suggested_next_reason` names which open requests exist
-  // when the suggestion ladder is empty but the substrate isn't.
-  const { root, cleanup } = bootstrap();
-  try {
-    // Alice authors a request naming bob as executor. Host has no
-    // role in the executor list, so actionableTransitions is empty
-    // for the host — but `status.pending.total: 1` would still show.
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['suggest'], { GUILD_ACTOR: 'human' });
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.suggested_next, null);
-    assert.ok(
-      typeof payload.suggested_next_reason === 'string' &&
-        payload.suggested_next_reason.includes('1 pending'),
-      `expected reason to mention 1 pending, got: ${payload.suggested_next_reason}`,
-    );
-    assert.match(payload.suggested_next_reason, /none names you as executor/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('suggest: self-wave pending is counted so the reason reconciles with queues', () => {
-  // dogfood 2026-05-29: an actor authored a self-wave pending (author ==
-  // executor). suggested_next stayed null (the ladder won't nudge
-  // self-approve — actionableTransitions requires from != actor for
-  // pending-as-executor), but the null-reason skipped it while boot's
-  // `queues: pending` counted it. The two surfaces disagreed (queues
-  // pending=2, reason said "1 pending"). The reason now carries a
-  // self-wave clause so the tallies match.
-  const { root, cleanup } = bootstrap();
-  try {
-    // alice: self-wave (author == executor). bob: a normal wave alice
-    // has no role in. Viewed as alice, neither is actionable, so
-    // suggested_next is null and both must be reflected in the reason.
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'solo', '--reason', 'r', '--executors', 'alice'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    runGate(
-      root,
-      ['request', '--from', 'bob', '--action', 'theirs', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'bob' },
-    );
-    const { stdout } = runGate(root, ['suggest'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.suggested_next, null);
-    const reason: string = payload.suggested_next_reason;
-    // The not-mine clause (bob's) AND the self-wave clause (alice's)
-    // both fire, so the pending tally reconciles with queues.pending=2.
-    assert.match(reason, /none names you as executor/);
-    assert.match(reason, /self-wave/);
-    // Reconciliation: boot's queues.pending must equal the count the
-    // reason accounts for. Both pending requests are acknowledged.
-    const boot = JSON.parse(
-      runGate(root, ['boot', '--format', 'json'], { GUILD_ACTOR: 'alice' }).stdout,
-    );
-    assert.equal(boot.status.pending.total, 2);
-    assert.equal((reason.match(/pending/g) ?? []).length >= 2, true,
-      `reason must account for both pending requests; got: ${reason}`);
-  } finally {
-    cleanup();
-  }
-});
-
-test('suggest: a lone self-wave pending no longer goes silent', () => {
-  // Regression for the silent-gap sub-case: when the ONLY open request
-  // is a self-wave pending, the pre-fix null-reason returned null
-  // (not-mine total was 0), so `gate suggest` said nothing while
-  // `queues: pending=1` showed one — exactly the "silence reads as a
-  // bug" failure the reason field exists to prevent.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'solo', '--reason', 'r', '--executors', 'alice'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['suggest'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.suggested_next, null);
-    assert.ok(
-      typeof payload.suggested_next_reason === 'string' &&
-        /self-wave/.test(payload.suggested_next_reason),
-      `lone self-wave pending must surface a reason; got: ${payload.suggested_next_reason}`,
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('suggest --format text: null + reason → (nothing urgent — <reason>)', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(
-      root,
-      ['suggest', '--format', 'text'],
-      { GUILD_ACTOR: 'human' },
-    );
-    assert.match(stdout, /^\(nothing urgent — /);
-    assert.match(stdout, /1 pending/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot: suggested_next_reason field is null when suggested_next is non-null', () => {
-  // When the suggest ladder picks a hint, the hint's own `reason`
-  // field carries the explanation — the sibling field stays null
-  // to avoid duplicate prose at the surface.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'bob' });
-    const payload = JSON.parse(stdout);
-    assert.ok(payload.suggested_next !== null);
-    assert.equal(payload.suggested_next_reason, null);
-  } finally {
-    cleanup();
-  }
-});
-
-test('suggest --format text: compact two-line form for humans', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(
-      root,
-      ['suggest', '--format', 'text'],
-      { GUILD_ACTOR: 'bob' },
-    );
-    // Line 1: → verb arg=val arg=val
-    // Line 2: the reason, indented
-    const lines = stdout.trim().split('\n');
-    assert.equal(lines.length, 2);
-    assert.match(lines[0]!, /^→ approve/);
-    assert.match(lines[0]!, /id=/);
-    assert.match(lines[0]!, /by=bob/);
-    assert.match(lines[1]!, /^  /);
-  } finally {
-    cleanup();
-  }
-});
-
-// ── boot.verbs_available_now: state-aware verb discovery ────────
-
-test('boot.verbs_available_now: actionable lists all valid transitions', () => {
-  // bob has executing-by-me (0001) AND unreviewed-mine (0002). suggested_next
-  // picks ONE; actionable lists ALL siblings so a branching agent sees
-  // the other options.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 't1', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    runGate(root, ['approve', rid(1), '--by', 'alice']);
-    runGate(root, ['execute', rid(1), '--by', 'bob']);
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 't2', '--reason', 'r', '--auto-review', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'bob' });
-    const p = JSON.parse(stdout);
-    const verbs = p.verbs_available_now.actionable.map(
-      (a: { verb: string }) => a.verb,
-    );
-    assert.ok(verbs.includes('complete'), 'complete missing');
-    assert.ok(verbs.includes('fail'), 'fail missing');
-    assert.ok(verbs.includes('review'), 'review missing');
-    // suggested_next is ONE of the actionable entries.
-    assert.ok(
-      verbs.includes(p.suggested_next.verb),
-      'suggested_next must be in actionable list',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot.verbs_available_now: always_readable present for anonymous caller', () => {
-  // Initial-agent discovery: without identity, actionable is empty
-  // (can't transition anything), but always_readable still names
-  // the read surface so newcomers see the map.
-  const { root, cleanup } = bootstrap();
-  try {
-    const { stdout } = runGate(root, ['boot']);
-    const p = JSON.parse(stdout);
-    assert.equal(p.verbs_available_now.actionable.length, 0);
-    assert.equal(p.verbs_available_now.requires_other_actor.length, 0);
-    assert.ok(p.verbs_available_now.always_readable.length >= 10);
-    assert.ok(p.verbs_available_now.always_readable.includes('suggest'));
-    assert.ok(p.verbs_available_now.always_readable.includes('schema'));
-    assert.ok(p.verbs_available_now.always_readable.includes('unresponded'));
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot.verbs_available_now: requires_other_actor surfaces pending blockers', () => {
-  // 2.A: a non-host author who filed a pending request sees the
-  // approval blocker — verb=approve, candidates=[host], reason
-  // explains why they can't act alone. This is the gap that bit
-  // first-time agents: suggested_next would return "by: host" with
-  // no obvious context for why the actor's own --by wouldn't work.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'pending', '--reason', 'r', '--executors', 'alice'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'alice' });
-    const p = JSON.parse(stdout);
-    assert.equal(p.verbs_available_now.actionable.length, 0);
-    assert.ok(
-      p.verbs_available_now.requires_other_actor.length >= 1,
-      'expected a pending-approval blocker for alice',
-    );
-    const blocker = p.verbs_available_now.requires_other_actor[0];
-    assert.equal(blocker.verb, 'approve');
-    assert.deepEqual(blocker.candidates, ['human']);
-    assert.match(blocker.reason, /pending/i);
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot.verbs_available_now: host self-approval doesnt double-list as blocker', () => {
-  // When the actor IS the host, pending requests on their own
-  // record show up under actionable (pending-as-executor) — NOT
-  // under requires_other_actor, since the host can self-approve.
-  const { root, cleanup } = bootstrap();
-  try {
-    // human is the host; have human file + name self executor
-    runGate(
-      root,
-      ['request', '--from', 'human', '--action', 'self', '--reason', 'r', '--executors', 'human'],
-      { GUILD_ACTOR: 'human' },
-    );
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'human' });
-    const p = JSON.parse(stdout);
-    assert.equal(
-      p.verbs_available_now.requires_other_actor.length,
-      0,
-      'host should not see their own pending-approval as blocker',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot.status.unresponded: counts unresponded concerns for the actor', () => {
-  // The orientation gap: an actor with concerns recorded against
-  // their completed work used to see status = all-zero, then
-  // separately have to remember to run `gate unresponded`. With
-  // the count surfaced in boot.status, a single boot call shows
-  // the concern queue exists. Same detector as `gate unresponded`,
-  // so the two surfaces never disagree.
-  const { root, cleanup } = bootstrap();
-  try {
-    // Register carol so she can review.
-    runGate(root, ['register', '--name', 'carol']);
-
-    // Lifecycle: alice files → human approves → bob executes → bob
-    // completes → carol records a devil/concern verdict.
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'rushed', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const id = rid(1);
-    runGate(root, ['approve', id, '--by', 'human']);
-    runGate(root, ['execute', id, '--by', 'bob']);
-    runGate(root, ['complete', id, '--by', 'bob']);
-    runGate(root, [
-      'review', id, '--by', 'carol', '--lense', 'devil',
-      '--verdict', 'concern', '--comment', 'rushed',
-    ]);
-
-    const { stdout: aliceBoot } = runGate(
-      root,
-      ['boot'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const alice = JSON.parse(aliceBoot);
-    assert.equal(
-      alice.status.unresponded,
-      1,
-      'alice (author) should see her unresponded concern in boot.status',
-    );
-
-    // Cross-check: gate status surfaces the same number.
-    const { stdout: aliceStatus } = runGate(
-      root,
-      ['status', '--format', 'json'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(JSON.parse(aliceStatus).unresponded, 1);
-
-    // bob (executor, not in authorship set) has no unresponded.
-    const { stdout: bobBoot } = runGate(
-      root,
-      ['boot'],
-      { GUILD_ACTOR: 'bob' },
-    );
-    assert.equal(JSON.parse(bobBoot).status.unresponded, 0);
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot.verbs_available_now: executor (≠author) doesnt double-list as blocker', () => {
-  // When the actor is a non-host executor named on a pending request
-  // they didn't author, the pending-as-executor predicate puts approve
-  // into actionable. Listing the same id+verb under requires_other_actor
-  // would contradict actionable for the same record — the gap a fresh
-  // agent reading boot would notice immediately.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'cross', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'bob' });
-    const p = JSON.parse(stdout);
-    const actionable = p.verbs_available_now.actionable;
-    const blockers = p.verbs_available_now.requires_other_actor;
-    assert.ok(
-      actionable.some(
-        (a: { verb: string; id: string }) =>
-          a.verb === 'approve' && a.id.startsWith('20'),
-      ),
-      'executor should see approve as actionable',
-    );
-    assert.equal(
-      blockers.length,
-      0,
-      'executor (≠ author) should not see their own approve as a blocker',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('write response suggested_next carries actor_resolved', () => {
-  // 2.E: the boolean lets an orchestrator branch without parsing
-  // `args.by` against the env. True when args.by is absent or
-  // matches GUILD_ACTOR, false otherwise.
-  const { root, cleanup } = bootstrap();
-  try {
-    const created = runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'alice', '--format', 'json'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const payload = JSON.parse(created.stdout);
-    // Pending state suggests approve by host (human). alice is not
-    // the host, so actor_resolved=false.
-    assert.equal(payload.suggested_next.verb, 'approve');
-    assert.equal(payload.suggested_next.args.by, 'human');
-    assert.equal(payload.suggested_next.actor_resolved, false);
-  } finally {
-    cleanup();
-  }
-});
-
-test('boot.verbs_available_now: actionable entries carry id + reason', () => {
-  // The reason converts "approve exists" into "approve is valid on
-  // this id because …" — teaching not just catalog.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['request', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'bob' });
-    const p = JSON.parse(stdout);
-    const approve = p.verbs_available_now.actionable.find(
-      (a: { verb: string }) => a.verb === 'approve',
-    );
-    assert.ok(approve, 'approve should be actionable');
-    assert.equal(approve.id, rid(1));
-    assert.match(approve.reason, /pending/);
-    assert.match(approve.reason, /executor/);
-  } finally {
-    cleanup();
-  }
-});
-
-// ── voice calibration: Two-Persona Devil gets a memory ──────────
-
-test('voices --with-calibration: aligns verdicts against terminal outcomes', () => {
-  // Carol files sharp concerns on things that later fail — her
-  // devil-lense calibration should read as "trusted". Bob rubber-
-  // stamps ok on things that fail — "learning".
-  const { root, cleanup } = bootstrap();
-  try {
-    // Register carol (only alice + bob exist in the bootstrap).
-    const r = runGate(root, ['register', '--name', 'carol']);
-    assert.equal(r.status, 0);
-
-    // 7 lifecycles: 4 failed + 3 completed. Bob always ok, carol
-    // rejects failures and oks completions.
-    const outcomes: Array<['completed' | 'failed', number]> = [
-      ['completed', 1], ['failed', 2], ['failed', 3], ['completed', 4],
-      ['failed', 5], ['failed', 6], ['completed', 7],
-    ];
-    for (const [state, n] of outcomes) {
-      const id = rid(n);
-      runGate(
-        root,
-        ['request', '--from', 'alice', '--action', `t${n}`, '--reason', 'r', '--executors', 'alice'],
-        { GUILD_ACTOR: 'alice' },
-      );
-      runGate(root, ['approve', id, '--by', 'alice']);
-      runGate(root, ['execute', id, '--by', 'alice']);
-      if (state === 'completed') {
-        runGate(root, ['complete', id, '--by', 'alice']);
-      } else {
-        runGate(root, ['fail', id, '--by', 'alice', '--reason', 'nope']);
-      }
-      runGate(root, [
-        'review', id, '--by', 'bob', '--lense', 'devil', '--verdict', 'ok', '--comment', 'b',
-      ]);
-      const carolV = state === 'completed' ? 'ok' : 'reject';
-      runGate(root, [
-        'review', id, '--by', 'carol', '--lense', 'devil', '--verdict', carolV, '--comment', 'c',
-      ]);
-    }
-
-    const { stdout: carolJson } = runGate(
-      root,
-      ['voices', 'carol', '--format', 'json', '--with-calibration'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const carol = JSON.parse(carolJson).calibration;
-    assert.equal(carol.by_lens.devil.status, 'trusted');
-    assert.equal(carol.by_lens.devil.aligned, 7);
-    assert.equal(carol.by_lens.devil.missed, 0);
-
-    const { stdout: bobJson } = runGate(
-      root,
-      ['voices', 'bob', '--format', 'json', '--with-calibration'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const bob = JSON.parse(bobJson).calibration;
-    assert.equal(bob.by_lens.devil.status, 'learning');
-    assert.equal(bob.by_lens.devil.aligned, 3);
-    assert.equal(bob.by_lens.devil.missed, 4);
-  } finally {
-    cleanup();
-  }
-});
-
-test('voices: self-view hides calibration (no self-optimisation)', () => {
-  // Voter shouldn't see their own score — the calibration only lands
-  // when viewing OTHER voices. Keeps the signal honest (if you can't
-  // see it, you can't game it).
-  const { root, cleanup } = bootstrap();
-  try {
-    // Seed one review so there's data to hide.
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    runGate(root, [
-      'review', rid(1), '--by', 'bob', '--lense', 'devil', '--verdict', 'ok', '--comment', 'b',
-    ]);
-
-    // bob views himself: calibration block NOT rendered in text.
-    const selfText = runGate(root, ['voices', 'bob', '--format', 'text'], { GUILD_ACTOR: 'bob' });
-    assert.equal(/calibration/.test(selfText.stdout), false);
-
-    // alice views bob: calibration block IS present (or at least the
-    // footer header renders, even if sample count is low).
-    const viewText = runGate(root, ['voices', 'bob', '--format', 'text'], { GUILD_ACTOR: 'alice' });
-    assert.match(viewText.stdout, /calibration/);
-
-    // JSON path: self-view returns null calibration under the flag.
-    const selfJson = runGate(
-      root,
-      ['voices', 'bob', '--format', 'json', '--with-calibration'],
-      { GUILD_ACTOR: 'bob' },
-    );
-    const payload = JSON.parse(selfJson.stdout);
-    assert.equal(payload.calibration, null);
-  } finally {
-    cleanup();
-  }
-});
-
-test('voices: default JSON shape unchanged (backward compat)', () => {
-  // `--with-calibration` is opt-in; without it, the JSON remains an
-  // array of utterances so existing consumers don't break.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout } = runGate(root, ['voices', 'alice', '--format', 'json']);
-    const payload = JSON.parse(stdout);
-    assert.ok(Array.isArray(payload), 'default JSON should still be an array');
-  } finally {
-    cleanup();
-  }
-});
-
-test('voices calibration: samples < 5 reads as "uncalibrated"', () => {
-  // Noise floor: a handful of verdicts isn't signal. Show the count
-  // so a reader sees where we are on the ramp, but don't claim a
-  // status from incomplete data.
-  const { root, cleanup } = bootstrap();
-  try {
-    for (let n = 1; n <= 3; n++) {
-      runGate(
-        root,
-        ['fast-track', '--from', 'alice', '--action', `t${n}`, '--reason', 'r'],
-        { GUILD_ACTOR: 'alice' },
-      );
-      runGate(root, [
-        'review', rid(n), '--by', 'bob', '--lense', 'devil', '--verdict', 'ok', '--comment', 'b',
-      ]);
-    }
-    const { stdout } = runGate(
-      root,
-      ['voices', 'bob', '--format', 'json', '--with-calibration'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const calib = JSON.parse(stdout).calibration;
-    assert.equal(calib.by_lens.devil.status, 'uncalibrated');
-    assert.equal(calib.by_lens.devil.sample_count, 3);
-    assert.equal(calib.by_lens.devil.alignment, null);
-  } finally {
-    cleanup();
-  }
-});
-
-test('voices calibration: verdict=concern + state=failed counts as aligned', () => {
-  // Pre-fix this case was excluded entirely — the source-code
-  // header documented `concern + failed → aligned` but the
-  // implementation had `// verdict === 'concern' intentionally
-  // excluded`. Reviewers who used `concern` as their primary
-  // signal got zero credit and registered as uncalibrated forever.
-  // This test pins the contract directly so the documented v1
-  // alignment rules can't quietly drift again.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(root, ['register', '--name', 'carol']);
-    // 5 cycles, all failures, carol flags `concern` on each. Pre-
-    // fix: 0 samples (uncalibrated). Post-fix: 5 aligned, trusted.
-    for (let n = 1; n <= 5; n++) {
-      const id = rid(n);
-      runGate(
-        root,
-        ['request', '--from', 'alice', '--action', `t${n}`, '--reason', 'r', '--executors', 'alice'],
-        { GUILD_ACTOR: 'alice' },
-      );
-      runGate(root, ['approve', id, '--by', 'alice']);
-      runGate(root, ['execute', id, '--by', 'alice']);
-      runGate(root, ['fail', id, '--by', 'alice', '--reason', 'nope']);
-      runGate(root, [
-        'review', id, '--by', 'carol',
-        '--lense', 'devil', '--verdict', 'concern', '--comment', 'flagged',
-      ]);
-    }
-    const { stdout } = runGate(
-      root,
-      ['voices', 'carol', '--format', 'json', '--with-calibration'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const calib = JSON.parse(stdout).calibration;
-    assert.equal(calib.by_lens.devil.aligned, 5);
-    assert.equal(calib.by_lens.devil.missed, 0);
-    assert.equal(calib.by_lens.devil.sample_count, 5);
-    assert.equal(calib.by_lens.devil.status, 'trusted');
-  } finally {
-    cleanup();
-  }
-});
-
-test('voices calibration: verdict=concern + state=completed stays excluded (soft)', () => {
-  // The other concern branch — concern+completed = "soft", neither
-  // a win nor a miss. The header rationale: the work landed, which
-  // is consistent with "concern was noted and addressed" AND with
-  // "concern was overblown" — counting it either way would bias
-  // the score. Excluded from sample_count entirely.
-  //
-  // Pre-fix this was already the de-facto behaviour (because ALL
-  // concerns were excluded). Post-fix, concern+failed counts as
-  // aligned but concern+completed must stay excluded — pin that
-  // boundary so a future "count all concerns somehow" refactor
-  // doesn't quietly change the soft semantic.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(root, ['register', '--name', 'carol']);
-    // 5 cycles, all completed, carol's concerns. Expected: 0 samples
-    // (all soft, all excluded), uncalibrated.
-    for (let n = 1; n <= 5; n++) {
-      const id = rid(n);
-      runGate(
-        root,
-        ['request', '--from', 'alice', '--action', `t${n}`, '--reason', 'r', '--executors', 'alice'],
-        { GUILD_ACTOR: 'alice' },
-      );
-      runGate(root, ['approve', id, '--by', 'alice']);
-      runGate(root, ['execute', id, '--by', 'alice']);
-      runGate(root, ['complete', id, '--by', 'alice']);
-      runGate(root, [
-        'review', id, '--by', 'carol',
-        '--lense', 'devil', '--verdict', 'concern', '--comment', 'noted',
-      ]);
-    }
-    const { stdout } = runGate(
-      root,
-      ['voices', 'carol', '--format', 'json', '--with-calibration'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const calib = JSON.parse(stdout).calibration;
-    // The bucket exists (it's created on first sight of any verdict
-    // for the lense) but soft verdicts don't increment aligned or
-    // missed — sample_count stays 0 and status reads as uncalibrated.
-    // The boundary we're pinning: soft does NOT bias the score.
-    const devil = calib.by_lens.devil;
-    assert.ok(devil, 'devil bucket should exist for the touched lense');
-    assert.equal(devil.aligned, 0);
-    assert.equal(devil.missed, 0);
-    assert.equal(devil.sample_count, 0);
-    assert.equal(devil.status, 'uncalibrated');
-  } finally {
-    cleanup();
-  }
-});
-
-// ── gate thank: cross-actor appreciation primitive ──────────────
-
-test('thank: records appreciation with by/to/reason on the request', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { status } = runGate(
-      root,
-      ['thank', 'bob', '--for', rid(1), '--reason', 'nice work'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(status, 0);
-    const { stdout } = runGate(root, ['show', rid(1), '--fields', 'thanks']);
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.thanks.length, 1);
-    assert.equal(payload.thanks[0].by, 'alice');
-    assert.equal(payload.thanks[0].to, 'bob');
-    assert.equal(payload.thanks[0].reason, 'nice work');
-    assert.equal(typeof payload.thanks[0].at, 'string');
-  } finally {
-    cleanup();
-  }
-});
-
-test('thank: does NOT affect state or reviews (orthogonal record)', () => {
-  // Critical invariant: thanks is NOT a state transition and does
-  // NOT feed calibration. Keep these concerns orthogonal.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    runGate(root, ['thank', 'alice', '--for', rid(1), '--reason', 'ty'], {
-      GUILD_ACTOR: 'alice',
-    });
-    const { stdout } = runGate(root, ['show', rid(1)]);
-    const p = JSON.parse(stdout);
-    assert.equal(p.state, 'completed');
-    assert.equal(p.reviews.length, 0);
-    assert.equal(p.thanks.length, 1);
-  } finally {
-    cleanup();
-  }
-});
-
-test('thank: reason is optional', () => {
-  // Most of the time the fact of the thank is the signal; a reason
-  // is a grace note. The schema doesn't require it.
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { status } = runGate(root, ['thank', 'bob', '--for', rid(1)], {
-      GUILD_ACTOR: 'alice',
-    });
-    assert.equal(status, 0);
-    const { stdout } = runGate(root, ['show', rid(1), '--fields', 'thanks']);
-    const thanks = JSON.parse(stdout).thanks;
-    assert.equal(thanks[0].reason, undefined);
-  } finally {
-    cleanup();
-  }
-});
-
-test('thank: self-thank emits stderr notice but succeeds', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { status, stderr } = runGate(
-      root,
-      ['thank', 'alice', '--for', rid(1), '--reason', 'past me'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(status, 0);
-    assert.match(stderr, /self-thank/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('thank: unknown --to actor fails with a validation error', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { status, stderr } = runGate(
-      root,
-      ['thank', 'ghost', '--for', rid(1)],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(status, 1);
-    assert.match(stderr, /ghost/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('thank --dry-run: preview without persist', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    runGate(
-      root,
-      ['fast-track', '--from', 'alice', '--action', 'x', '--reason', 'r', '--executors', 'bob'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const { stdout, status } = runGate(
-      root,
-      ['thank', 'bob', '--for', rid(1), '--reason', 'preview', '--dry-run'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(status, 0);
-    const p = JSON.parse(stdout);
-    assert.equal(p.dry_run, true);
-    assert.equal(p.verb, 'thank');
-    assert.equal(p.preview.thanks.length, 1);
-    // Real record has no thanks yet.
-    const after = JSON.parse(
-      runGate(root, ['show', rid(1)]).stdout,
-    );
-    assert.ok(after.thanks === undefined || after.thanks.length === 0);
   } finally {
     cleanup();
   }
